requirements.txt
torch>=2.1.0
transformers>=4.38.0
datasets>=2.16.0
sentencepiece>=0.1.99
sacrebleu>=2.4.0
seqeval>=1.2.2
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
